server/requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.92.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-vision==3.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.26.0
imutils==0.5.4
lazy_loader==0.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
